refactor aws ec2 deployment
