+++ bentoml/yatai/deployment/aws_ec2/operator.py
@@ -3,12 +3,15 @@ import boto3
 import base64
 import json
 import logging
-from botocore.exceptions import ClientError
 
 from bentoml.utils.tempdir import TempDirectory
 
 from bentoml.utils.lazy_loader import LazyLoader
 from bentoml.saved_bundle import loader
+from bentoml.yatai.deployment.aws_ec2.templates import (
+    USER_INIT_SCRIPT_TEMPLATE,
+    CLOUDFORMATION_CONFIG_TEMPLATE,
+)
 from bentoml.yatai.deployment.operator import DeploymentOperatorBase
 from bentoml.yatai.proto.deployment_pb2 import (
     ApplyDeploymentResponse,
@@ -31,8 +34,10 @@ from bentoml.yatai.deployment.aws_utils import (
     delete_ecr_repository,
     get_scaling_group_instance_addresses,
     get_aws_user_id,
+    get_ecr_login_info,
+    create_ecr_repository_if_not_exists,
+    create_and_push_docker_image_to_ecr, call_sam_command,
 )
-from bentoml.utils.docker_utils import containerize_bento_service
 from bentoml.exceptions import (
     BentoMLException,
     InvalidArgument,
@@ -40,13 +45,6 @@ from bentoml.exceptions import (
 )
 from bentoml.yatai.proto.repository_pb2 import GetBentoRequest, BentoUri
 from bentoml.yatai.proto import status_pb2
-from bentoml.yatai.deployment.aws_ec2.utils import (
-    build_template,
-    package_template,
-    deploy_template,
-    get_endpoints_from_instance_address,
-    get_healthy_target,
-)
 from bentoml.yatai.deployment.aws_ec2.constants import (
     BENTOSERVICE_PORT,
     TARGET_HEALTH_CHECK_INTERVAL,
@@ -62,57 +60,23 @@ yatai_proto = LazyLoader("yatai_proto", globals(), "bentoml.yatai.proto")
 SAM_TEMPLATE_NAME = "template.yml"
 
 
-def _create_ecr_repo(repo_name, region):
-    """
-    Create ecr repository,in given region
-    args:
-        repo_name: repository name to create
-        region: aws region
-    """
-    try:
-        ecr_client = boto3.client("ecr", region)
-        repository = ecr_client.create_repository(
-            repositoryName=repo_name, imageScanningConfiguration={"scanOnPush": False}
-        )
-        registry_id = repository["repository"]["registryId"]
+def _generate_ec2_resources_names(namespace, name):
+    sam_template_name = generate_aws_compatible_string(
+        "btml-template-{namespace}-{name}".format(namespace=namespace, name=name)
+    )
 
-    except ecr_client.exceptions.RepositoryAlreadyExistsException:
-        all_repositories = ecr_client.describe_repositories(repositoryNames=[repo_name])
-        registry_id = all_repositories["repositories"][0]["registryId"]
+    deployment_stack_name = generate_aws_compatible_string(
+        "btml-stack-{namespace}-{name}".format(namespace=namespace, name=name)
+    )
 
-    return registry_id
+    repository_name = generate_aws_compatible_string(
+        "btml-repo-{namespace}-{name}".format(namespace=namespace, name=name)
+    )
 
-
-def _get_ecr_password(registry_id, region):
-    """
-    Get authentication token for registry to authenticate docker agent with ecr.
-    """
-    ecr_client = boto3.client("ecr", region)
-    try:
-        token_data = ecr_client.get_authorization_token(registryIds=[registry_id])
-        token = token_data["authorizationData"][0]["authorizationToken"]
-        registry_endpoint = token_data["authorizationData"][0]["proxyEndpoint"]
-        return token, registry_endpoint
-
-    except ClientError as error:
-        if (
-            error.response
-            and error.response["Error"]["Code"] == "InvalidParameterException"
-        ):
-            raise BentoMLException(
-                "Could not get token for registry {registry_id},{error}".format(
-                    registry_id=registry_id, error=error.response["Error"]["Message"]
-                )
-            )
-
-
-def _get_creds_from_token(token):
-    """
-    Decode ecr token into username and password.
-    """
-    cred_string = base64.b64decode(token).decode("ascii")
-    username, password = str(cred_string).split(":")
-    return username, password
+    elb_name = generate_aws_compatible_string(
+        "{namespace}-{name}".format(namespace=namespace, name=name), max_length=32,
+    )
+    return sam_template_name, deployment_stack_name, repository_name, elb_name
 
 
 def _make_user_data(registry, tag, region):
@@ -124,32 +88,10 @@ def _make_user_data(registry, tag, region):
         tag: bento tag
         region: AWS region
     """
-
-    base_format = """MIME-Version: 1.0
-Content-Type: multipart/mixed; boundary=\"==MYBOUNDARY==\"
-
---==MYBOUNDARY==
-Content-Type: text/cloud-config; charset=\"us-ascii\"
-
-runcmd:
-
-- sudo yum update -y
-- sudo amazon-linux-extras install docker -y
-- sudo service docker start
-- sudo usermod -a -G docker ec2-user
-- curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
-- unzip awscliv2.zip
-- sudo ./aws/install
-- ln -s /usr/bin/aws aws
-- aws ecr get-login-password --region {region}|docker login --username AWS --password-stdin {registry}
-- docker pull {tag}
-- docker run -p {bentoservice_port}:{bentoservice_port} {tag}
-
---==MYBOUNDARY==--
-""".format(  # noqa: E501
+    user_init_script = USER_INIT_SCRIPT_TEMPLATE.format(
         registry=registry, tag=tag, region=region, bentoservice_port=BENTOSERVICE_PORT
     )
-    encoded = base64.b64encode(base_format.encode("ascii")).decode("ascii")
+    encoded = base64.b64encode(user_init_script.encode("ascii")).decode("ascii")
     return encoded
 
 
@@ -186,226 +128,7 @@ def _make_cloudformation_template(
     template_file_path = os.path.join(project_dir, sam_template_name)
     with open(template_file_path, "a") as f:
         f.write(
-            """\
-AWSTemplateFormatVersion: 2010-09-09
-Transform: AWS::Serverless-2016-10-31
-Description: BentoML load balanced template
-Parameters:
-    AmazonLinux2LatestAmiId:
-        Type: AWS::SSM::Parameter::Value<AWS::EC2::Image::Id>
-        Default: {ami_id}
-Resources:
-    SecurityGroupResource:
-        Type: AWS::EC2::SecurityGroup
-        Properties:
-            GroupDescription: "security group for bentoservice"
-            SecurityGroupIngress:
-                -
-                    IpProtocol: tcp
-                    CidrIp: 0.0.0.0/0
-                    FromPort: 5000
-                    ToPort: 5000
-                -
-                    IpProtocol: tcp
-                    CidrIp: 0.0.0.0/0
-                    FromPort: 22
-                    ToPort: 22
-            VpcId: !Ref Vpc1
-
-    Ec2InstanceECRProfile:
-        Type: AWS::IAM::InstanceProfile
-        Properties:
-            Path: /
-            Roles: [!Ref EC2Role]
-
-    EC2Role:
-        Type: AWS::IAM::Role
-        Properties:
-            AssumeRolePolicyDocument:
-                Statement:
-                    -   Effect: Allow
-                        Principal:
-                            Service: [ec2.amazonaws.com]
-                        Action: ['sts:AssumeRole']
-            Path: /
-            Policies:
-                -   PolicyName: ecs-service
-                    PolicyDocument:
-                        Statement:
-                            -   Effect: Allow
-                                Action:
-                                    -   'ecr:GetAuthorizationToken'
-                                    -   'ecr:BatchGetImage'
-                                    -   'ecr:GetDownloadUrlForLayer'
-                                Resource: '*'
-
-    LaunchTemplateResource:
-        Type: AWS::EC2::LaunchTemplate
-        Properties:
-            LaunchTemplateName: {template_name}
-            LaunchTemplateData:
-                IamInstanceProfile:
-                    Arn: !GetAtt Ec2InstanceECRProfile.Arn
-                ImageId: !Ref AmazonLinux2LatestAmiId
-                InstanceType: {instance_type}
-                UserData: "{user_data}"
-                SecurityGroupIds:
-                - !GetAtt SecurityGroupResource.GroupId
-
-    TargetGroup:
-        Type: AWS::ElasticLoadBalancingV2::TargetGroup
-        Properties:
-            VpcId: !Ref Vpc1
-            Protocol: HTTP
-            Port: 5000
-            TargetType: instance
-            HealthCheckEnabled: true
-            HealthCheckIntervalSeconds: {target_health_check_interval_seconds}
-            HealthCheckPath: {target_health_check_path}
-            HealthCheckPort: {target_health_check_port}
-            HealthCheckProtocol: HTTP
-            HealthCheckTimeoutSeconds: {target_health_check_timeout_seconds}
-            HealthyThresholdCount: {target_health_check_threshold_count}
-
-    LoadBalancerSecurityGroup:
-        Type: AWS::EC2::SecurityGroup
-        Properties:
-            GroupDescription: "security group for loadbalancing"
-            VpcId: !Ref Vpc1
-            SecurityGroupIngress:
-                -
-                    IpProtocol: tcp
-                    CidrIp: 0.0.0.0/0
-                    FromPort: 80
-                    ToPort: 80
-
-    InternetGateway:
-        Type: AWS::EC2::InternetGateway
-
-    Gateway:
-        Type: AWS::EC2::VPCGatewayAttachment
-        Properties:
-            InternetGatewayId: !Ref InternetGateway
-            VpcId: !Ref Vpc1
-
-    PublicRouteTable:
-        Type: AWS::EC2::RouteTable
-        Properties:
-            VpcId: !Ref Vpc1
-
-    PublicRoute:
-        Type: AWS::EC2::Route
-        DependsOn: Gateway
-        Properties:
-            DestinationCidrBlock: 0.0.0.0/0
-            GatewayId: !Ref InternetGateway
-            RouteTableId: !Ref PublicRouteTable
-
-    RouteTableSubnetTwoAssociationOne:
-        Type: AWS::EC2::SubnetRouteTableAssociation
-        Properties:
-          RouteTableId: !Ref PublicRouteTable
-          SubnetId: !Ref Subnet1
-    RouteTableSubnetTwoAssociationTwo:
-        Type: AWS::EC2::SubnetRouteTableAssociation
-        Properties:
-          RouteTableId: !Ref PublicRouteTable
-          SubnetId: !Ref Subnet2
-
-    Vpc1:
-        Type: AWS::EC2::VPC
-        Properties:
-            CidrBlock: 172.31.0.0/16
-            EnableDnsHostnames: true
-            EnableDnsSupport: true
-            InstanceTenancy: default
-
-    Subnet1:
-        Type: AWS::EC2::Subnet
-        Properties:
-            VpcId: !Ref Vpc1
-            AvailabilityZone:
-                Fn::Select:
-                    - 0
-                    - Fn::GetAZs: ""
-            CidrBlock: 172.31.16.0/20
-            MapPublicIpOnLaunch: true
-
-    Subnet2:
-        Type: AWS::EC2::Subnet
-        Properties:
-            VpcId: !Ref Vpc1
-            AvailabilityZone:
-                Fn::Select:
-                    - 1
-                    - Fn::GetAZs: ""
-            CidrBlock: 172.31.0.0/20
-            MapPublicIpOnLaunch: true
-
-    LoadBalancer:
-        Type: AWS::ElasticLoadBalancingV2::LoadBalancer
-        Properties:
-            IpAddressType: ipv4
-            Name: {elb_name}
-            Scheme: internet-facing
-            SecurityGroups:
-                - !Ref LoadBalancerSecurityGroup
-            Subnets:
-                - !Ref Subnet1
-                - !Ref Subnet2
-            Type: application
-
-    Listener:
-        Type: AWS::ElasticLoadBalancingV2::Listener
-        Properties:
-            DefaultActions:
-                -   Type: forward
-                    TargetGroupArn: !Ref TargetGroup
-            LoadBalancerArn: !Ref LoadBalancer
-            Port: 80
-            Protocol: HTTP
-
-    AutoScalingGroup:
-        Type: AWS::AutoScaling::AutoScalingGroup
-        DependsOn: Gateway
-        Properties:
-            MinSize: {autoscaling_min_size}
-            MaxSize: {autoscaling_max_size}
-            DesiredCapacity: {autoscaling_desired_capacity}
-            AvailabilityZones:
-                - Fn::Select:
-                    - 0
-                    - Fn::GetAZs: ""
-                - Fn::Select:
-                    - 1
-                    - Fn::GetAZs: ""
-            LaunchTemplate:
-                LaunchTemplateId: !Ref LaunchTemplateResource
-                Version: !GetAtt LaunchTemplateResource.LatestVersionNumber
-            TargetGroupARNs:
-                - !Ref TargetGroup
-            VPCZoneIdentifier:
-            - !Ref Subnet1
-            - !Ref Subnet2
-        UpdatePolicy:
-            AutoScalingReplacingUpdate:
-                WillReplace: true
-
-Outputs:
-    S3Bucket:
-        Value: {s3_bucket_name}
-        Description: Bucket to store sam artifacts
-    AutoScalingGroup:
-        Value: !Ref AutoScalingGroup
-        Description: Autoscaling group name
-    TargetGroup:
-        Value: !Ref TargetGroup
-        Description: Target group for load balancer
-    Url:
-        Value: !Join ['', ['http://', !GetAtt [LoadBalancer, DNSName]]]
-        Description: URL of the bento service
-
-""".format(
+            CLOUDFORMATION_CONFIG_TEMPLATE.format(
                 ami_id=ami_id,
                 template_name=sam_template_name,
                 instance_type=instance_type,
@@ -425,6 +148,82 @@ Outputs:
     return template_file_path
 
 
+def build_template(template_file_path, project_directory, region):
+    status_code, stdout, stderr = call_sam_command(
+        ["build", "-t", template_file_path], project_directory, region
+    )
+
+    if status_code != 0:
+        error_message = stderr if stderr else stdout
+        raise BentoMLException("Failed to build ec2 service {}".format(error_message))
+
+    return status_code, stdout, stderr
+
+
+def package_template(s3_bucket_name, project_directory, region):
+    status_code, stdout, stderr = call_sam_command(
+        [
+            "package",
+            "--output-template-file",
+            "packaged.yaml",
+            "--s3-bucket",
+            s3_bucket_name,
+        ],
+        project_directory,
+        region,
+    )
+    if status_code != 0:
+        error_message = stderr if stderr else stdout
+        raise BentoMLException("Failed to package ec2 service {}".format(error_message))
+    return status_code, stdout, stderr
+
+
+def deploy_template(stack_name, s3_bucket_name, project_directory, region):
+    status_code, stdout, stderr = call_sam_command(
+        [
+            "deploy",
+            "--template-file",
+            "packaged.yaml",
+            "--stack-name",
+            stack_name,
+            "--capabilities",
+            "CAPABILITY_IAM",
+            "--s3-bucket",
+            s3_bucket_name,
+        ],
+        project_directory,
+        region,
+    )
+    if status_code != 0:
+        error_message = stderr if stderr else stdout
+        raise BentoMLException("Failed to deploy ec2 service {}".format(error_message))
+    return status_code, stdout, stderr
+
+
+def get_endpoints_from_instance_address(instances, api_names):
+    all_endpoints = []
+    for instance in instances:
+        if instance["state"] == AWS_EC2_IN_SERVICE_STATE:
+            for api in api_names:
+                all_endpoints.append(
+                    "{ep}:{port}/{api}".format(
+                        ep=instance["endpoint"], port=BENTOSERVICE_PORT, api=api
+                    )
+                )
+
+    return all_endpoints
+
+
+def get_healthy_target(target_group_arn, region):
+    eb_client = boto3.client("elbv2", region)
+
+    all_targets = eb_client.describe_target_health(TargetGroupArn=target_group_arn)
+    for instance in all_targets["TargetHealthDescriptions"]:
+        if instance["TargetHealth"]["State"] == TARGET_HEALTHY_STATUS:
+            return instance
+    return None
+
+
 class AwsEc2DeploymentOperator(DeploymentOperatorBase):
     def deploy_service(
         self,
@@ -435,50 +234,27 @@ class AwsEc2DeploymentOperator(DeploymentOperatorBase):
         s3_bucket_name,
         region,
     ):
-        sam_template_name = generate_aws_compatible_string(
-            "btml-template-{namespace}-{name}".format(
-                namespace=deployment_pb.namespace, name=deployment_pb.name
-            )
-        )
-
-        deployment_stack_name = generate_aws_compatible_string(
-            "btml-stack-{namespace}-{name}".format(
-                namespace=deployment_pb.namespace, name=deployment_pb.name
-            )
-        )
-
-        repo_name = generate_aws_compatible_string(
-            "btml-repo-{namespace}-{name}".format(
-                namespace=deployment_pb.namespace, name=deployment_pb.name
-            )
-        )
-
-        elb_name = generate_aws_compatible_string(
-            "{namespace}-{name}".format(
-                namespace=deployment_pb.namespace, name=deployment_pb.name
-            ),
-            max_length=32,
-        )
+        (
+            sam_template_name,
+            deployment_stack_name,
+            repository_name,
+            elb_name,
+        ) = _generate_ec2_resources_names(deployment_pb.namespace, deployment_pb.name)
 
         with TempDirectory() as project_path:
-            registry_id = _create_ecr_repo(repo_name, region)
-            registry_token, registry_url = _get_ecr_password(registry_id, region)
-            registry_username, registry_password = _get_creds_from_token(registry_token)
+            registry_url, username, password = get_ecr_login_info(region)
+            create_ecr_repository_if_not_exists(region, repository_name)
 
             registry_domain = registry_url.replace("https://", "")
-            push_tag = f"{registry_domain}/{repo_name}"
+            push_tag = f"{registry_domain}/{repository_name}"
             pull_tag = push_tag + f":{deployment_spec.bento_version}"
 
-            logger.info("Containerizing service")
-            containerize_bento_service(
-                bento_name=deployment_spec.bento_name,
-                bento_version=deployment_spec.bento_version,
-                saved_bundle_path=bento_path,
-                push=True,
-                tag=push_tag,
-                build_arg={},
-                username=registry_username,
-                password=registry_password,
+            create_and_push_docker_image_to_ecr(
+                region=region,
+                repository_name=repository_name,
+                docker_context_path=bento_path,
+                dockerfile_name='Dockerfile',
+                image_tag='',
             )
 
             logger.info("Generating user data")
@@ -501,17 +277,13 @@ class AwsEc2DeploymentOperator(DeploymentOperatorBase):
                 sam_template_name, aws_ec2_deployment_config.region, project_path
             )
 
-            logger.info("Building service")
             build_template(
                 template_file_path, project_path, aws_ec2_deployment_config.region
             )
 
-            logger.info("Packaging service")
             package_template(
                 s3_bucket_name, project_path, aws_ec2_deployment_config.region
             )
-
-            logger.info("Deploying service")
             deploy_template(
                 deployment_stack_name,
                 s3_bucket_name,
